add the auth component
