+++ angular/src/app/app.component.html
@@ -1,3 +1,1 @@
-<div class="container ">
-  <router-outlet></router-outlet>
-</div>
+<router-outlet></router-outlet>
