+++ angular/src/app/app.component.scss
@@ -1,4 +1,4 @@
-.container {
+:host {
   align-items: center;
   display: flex;
   height: 100vh;
