+++ angular/src/app/screens/auth/auth-routing.module.ts
@@ -1,10 +1,11 @@
 import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { BooksComponent } from './books.component';
+import { Routes, RouterModule } from '@angular/router';
+import { AuthComponent } from './auth.component';
+
 const routes: Routes = [
   {
     path: '',
-    component: BooksComponent,
+    component: AuthComponent,
   },
 ];
 
@@ -12,4 +13,4 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule],
 })
-export class BooksRoutingModule {}
+export class AuthRoutingModule {}
