+++ angular/src/app/screens/auth/auth.component.scss
@@ -4,9 +4,3 @@
   height: 100vh;
   width: 100%;
 }
-
-.container {
-  background-color: $wild-sand;
-  height: 100vh;
-  width: 100%;
-}
