+++ angular/src/app/screens/auth/auth.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AuthComponent } from './auth.component';
+import { AuthRoutingModule } from './auth-routing.module';
+import { NavbarComponent } from 'src/app/components/navbar/navbar.component';
+
+@NgModule({
+  declarations: [AuthComponent, NavbarComponent],
+  imports: [CommonModule, AuthRoutingModule],
+})
+export class AuthModule {}
