--- angular/src/app/screens/books/books.component.ts
@@ -1,8 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-books',
-  templateUrl: './books.component.html',
-  styleUrls: ['./books.component.scss'],
-})
-export class BooksComponent {}
